Move duration histogram into a submodule of apollo metrics.
