+++ apollo-router/src/plugins/telemetry/metrics/apollo.rs
@@ -9,6 +9,7 @@ use apollo_spaceport::{
 };
 use async_trait::async_trait;
 use deadpool::{managed, Runtime};
+use duration_histogram::DurationHistogram;
 use futures::channel::mpsc;
 use futures_batch::ChunksTimeoutStreamExt;
 use serde::Serialize;
@@ -17,6 +18,8 @@ use std::ops::AddAssign;
 use std::time::{Duration, SystemTime};
 use tower::BoxError;
 use url::Url;
+
+mod duration_histogram;
 
 const DEFAULT_BATCH_SIZE: usize = 65_536;
 const DEFAULT_QUEUE_SIZE: usize = 65_536;
@@ -468,67 +471,6 @@ impl managed::Manager for ReporterManager {
     }
 }
 
-#[derive(Serialize)]
-struct DurationHistogram {
-    buckets: Vec<i64>,
-    entries: u64,
-}
-
-impl Default for DurationHistogram {
-    fn default() -> Self {
-        DurationHistogram::new(None)
-    }
-}
-
-// The TS implementation of DurationHistogram does Run Length Encoding (RLE)
-// to replace sequences of empty buckets with negative numbers. This
-// implementation doesn't because:
-// Spending too much time in the export() fn exerts back-pressure into the
-// telemetry framework and leads to dropped data spans. Given that the
-// histogram data is ultimately gzipped for transfer, I wasn't entirely
-// sure that this extra processing was worth performing.
-impl DurationHistogram {
-    const DEFAULT_SIZE: usize = 74; // Taken from TS implementation
-    const MAXIMUM_SIZE: usize = 383; // Taken from TS implementation
-    const EXPONENT_LOG: f64 = 0.09531017980432493f64; // ln(1.1) Update when ln() is a const fn (see: https://github.com/rust-lang/rust/issues/57241)
-    fn new(init_size: Option<usize>) -> Self {
-        Self {
-            buckets: vec![0; init_size.unwrap_or(DurationHistogram::DEFAULT_SIZE)],
-            entries: 0,
-        }
-    }
-
-    fn duration_to_bucket(duration: Duration) -> usize {
-        // If you use as_micros() here to avoid the divide, tests will fail
-        // Because, internally, as_micros() is losing remainders
-        let log_duration = f64::ln(duration.as_nanos() as f64 / 1000.0);
-        let unbounded_bucket = f64::ceil(log_duration / DurationHistogram::EXPONENT_LOG);
-
-        if unbounded_bucket.is_nan() || unbounded_bucket <= 0f64 {
-            return 0;
-        } else if unbounded_bucket > DurationHistogram::MAXIMUM_SIZE as f64 {
-            return DurationHistogram::MAXIMUM_SIZE;
-        }
-
-        unbounded_bucket as usize
-    }
-
-    fn increment_duration(&mut self, duration: Duration, value: i64) {
-        self.increment_bucket(DurationHistogram::duration_to_bucket(duration), value)
-    }
-
-    fn increment_bucket(&mut self, bucket: usize, value: i64) {
-        if bucket > DurationHistogram::MAXIMUM_SIZE {
-            panic!("bucket is out of bounds of the bucket array");
-        }
-        self.entries += value as u64;
-        if bucket >= self.buckets.len() {
-            self.buckets.resize(bucket + 1, 0);
-        }
-        self.buckets[bucket] += value;
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::future::Future;
@@ -544,123 +486,6 @@ mod test {
 
     use super::super::super::config;
     use super::*;
-
-    // DurationHistogram Tests
-    impl DurationHistogram {
-        fn to_array(&self) -> Vec<i64> {
-            let mut result = vec![];
-            let mut buffered_zeroes = 0;
-
-            for value in &self.buckets {
-                if *value == 0 {
-                    buffered_zeroes += 1;
-                } else {
-                    if buffered_zeroes == 1 {
-                        result.push(0);
-                    } else if buffered_zeroes != 0 {
-                        result.push(0 - buffered_zeroes);
-                    }
-                    result.push(*value);
-                    buffered_zeroes = 0;
-                }
-            }
-            result
-        }
-    }
-
-    #[test]
-    fn it_generates_empty_histogram() {
-        let histogram = DurationHistogram::new(None);
-        let expected: Vec<i64> = vec![];
-        assert_eq!(histogram.to_array(), expected);
-    }
-
-    #[test]
-    fn it_generates_populated_histogram() {
-        let mut histogram = DurationHistogram::new(None);
-        histogram.increment_bucket(100, 1);
-        assert_eq!(histogram.to_array(), vec![-100, 1]);
-        histogram.increment_bucket(102, 1);
-        assert_eq!(histogram.to_array(), vec![-100, 1, 0, 1]);
-        histogram.increment_bucket(382, 1);
-        assert_eq!(histogram.to_array(), vec![-100, 1, 0, 1, -279, 1]);
-    }
-
-    #[test]
-    fn it_buckets_to_zero_and_one() {
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(0)),
-            0
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1)),
-            0
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(999)),
-            0
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1000)),
-            0
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1001)),
-            1
-        );
-    }
-
-    #[test]
-    fn it_buckets_to_one_and_two() {
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1100)),
-            1
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1101)),
-            2
-        );
-    }
-
-    #[test]
-    fn it_buckets_to_threshold() {
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(10000)),
-            25
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(10834)),
-            25
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(10835)),
-            26
-        );
-    }
-
-    #[test]
-    fn it_buckets_common_times() {
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1e5 as u64)),
-            49
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1e6 as u64)),
-            73
-        );
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1e9 as u64)),
-            145
-        );
-    }
-
-    #[test]
-    fn it_limits_to_last_bucket() {
-        assert_eq!(
-            DurationHistogram::duration_to_bucket(Duration::from_nanos(1e64 as u64)),
-            DurationHistogram::MAXIMUM_SIZE
-        );
-    }
 
     #[test]
     fn test_aggregation() {
